organizando logica de negocio,criando services e dtos

+++ alura/src/main/java/src/main/java/br/com/alura/modelo/Manager.java
@@ -6,7 +6,7 @@ import src.main.java.br.com.alura.enums.UserRole;
 
 public class Manager extends User{
 
-	public Manager(int id, String firstName, String lastName, String email, UserRole role ,LocalDateTime updated_at) {
+	public Manager(Long id, String firstName, String lastName, String email, UserRole role ,LocalDateTime updated_at) {
 
 
 		User user =  User.builder()
